Move some data shaping code from the vaccinations page to a hook
