+++ packages/app/src/pages/landelijk/vaccinaties.tsx
@@ -28,6 +28,7 @@ import { VisuallyHidden } from '~/components-styled/visually-hidden';
 import { FCWithLayout } from '~/domain/layout/layout';
 import { getNationalLayout } from '~/domain/layout/national-layout';
 import { createDeliveryTooltipFormatter } from '~/domain/vaccines/create-delivery-tooltip-formatter';
+import { useVaccineDeliveryData } from '~/domain/vaccines/use-vaccine-delivery-data';
 import { useVaccineNames } from '~/domain/vaccines/use-vaccine-names';
 import siteText from '~/locale/index';
 import { createPageArticlesQuery } from '~/queries/create-page-articles-query';
@@ -65,21 +66,16 @@ const VaccinationPage: FCWithLayout<typeof getStaticProps> = ({
   const additions = text.expected_page_additions.additions.filter(
     (x) => x.length
   );
-  const vaccineDeliveryValues = [...data.vaccine_delivery.values];
-  const vaccineDeliveryEstimateValues = [
-    ...data.vaccine_delivery_estimate.values,
-  ];
-  const vaccineAdministeredValues = [...data.vaccine_administered.values];
-  const vaccineAdministeredEstimateValues = [
-    ...data.vaccine_administered_estimate.values,
-  ];
 
   const vaccineNames = useVaccineNames(data.vaccine_administered.last_value);
 
-  // add the first estimate to the delivered values, otherwise the lines and stacks will
-  // have a gap between them
-  vaccineDeliveryValues.push({ ...vaccineDeliveryEstimateValues[0] });
-  vaccineAdministeredValues.push({ ...vaccineAdministeredEstimateValues[0] });
+  const [
+    vaccineDeliveryValues,
+    vaccineDeliveryEstimateValues,
+    vaccineAdministeredValues,
+    vaccineAdministeredEstimateValues,
+  ] = useVaccineDeliveryData(data);
+
   return (
     <>
       <SEOHead
